refactor: isCommit redefined as type guard
