+++ src/renderer/src/automerge/index.ts
@@ -7,7 +7,17 @@ export type VersionedDocument = {
   content: Automerge.Doc<string>;
 };
 
-export const isCommit = (change: Automerge.DecodedChange) => {
+// Commit is a special type of an (automerge) change that
+// strictly has a message and a time
+export type Commit = {
+  hash: string;
+  message: string;
+  time: Date;
+};
+// this is a TS type guard to check if a change is a commit
+export const isCommit = (
+  change: Automerge.DecodedChange | Commit
+): change is Commit => {
   // we make the rules!
   return Boolean(change.message && change.time);
 };
